[FEAT] : 옷 타입별 버튼 기능 추가(안쓸 수도 있음)
